add stream chunk downloading as option
