+++ statement.go
@@ -13,9 +13,11 @@ const (
 	// MultiStatementCount controls the number of queries to execute in a single API call
 	MultiStatementCount paramKey = "MULTI_STATEMENT_COUNT"
 	// AsyncMode tells the server to not block the request on executing the entire query
-	AsyncMode paramKey = "ASYNC_MODE_QUERY"
+	asyncMode paramKey = "ASYNC_MODE_QUERY"
 	// QueryIDChan is the channel to receive the query ID from
 	QueryIDChan paramKey = "QUERY_ID_CHANNEL"
+	// streamChunkDownload determines whether to use a stream based chunk downloader
+	streamChunkDownload paramKey = "STREAM_CHUNK_DOWNLOAD"
 )
 
 type snowflakeStmt struct {
@@ -61,11 +63,16 @@ func WithMultiStatement(ctx context.Context, num int) (context.Context, error) {
 }
 
 // WithAsyncMode returns a context that allows execution of query in async mode
-func WithAsyncMode(ctx context.Context) (context.Context, error) {
-	return context.WithValue(ctx, AsyncMode, true), nil
+func WithAsyncMode(ctx context.Context) context.Context {
+	return context.WithValue(ctx, asyncMode, true)
 }
 
 // WithQueryIDChan returns a context that contains the channel to receive the query ID
 func WithQueryIDChan(ctx context.Context, c chan<- string) context.Context {
 	return context.WithValue(ctx, QueryIDChan, c)
+}
+
+// WithStreamDownloader returns a context that allows the use of a stream based chunk downloader
+func WithStreamDownloader(ctx context.Context) context.Context {
+	return context.WithValue(ctx, streamChunkDownload, true)
 }
